Restructured oauth2 module and improved readability
Based on review feedback.
